+++ seqr/utils/search/hail_search_utils.py
@@ -2,7 +2,7 @@ from collections import defaultdict
 from django.db.models import F, Min
 
 import requests
-from reference_data.models import Omim, GeneConstraint, GENOME_VERSION_LOOKUP
+from reference_data.models import Omim, GeneConstraint, GENOME_VERSION_LOOKUP, GENOME_VERSION_GRCh38
 from seqr.models import Sample, PhenotypePrioritization
 from seqr.utils.search.constants import PRIORITIZED_GENE_SORT, X_LINKED_RECESSIVE
 from seqr.utils.xpos_utils import MIN_POS, MAX_POS
@@ -13,11 +13,12 @@ def _hail_backend_url(path):
     return f'{HAIL_BACKEND_SERVICE_HOSTNAME}:{HAIL_BACKEND_SERVICE_PORT}/{path}'
 
 
-def _execute_search(search_body, user, path='search'):
+def _execute_search(search_body, user, path='search', exception_map=None):
     response = requests.post(_hail_backend_url(path), json=search_body, headers={'From': user.email}, timeout=300)
 
     if response.status_code >= 400:
-        raise requests.HTTPError(response.text or response.reason, response=response)
+        error = (exception_map or {}).get(response.status_code) or response.text or response.reason
+        raise requests.HTTPError(error, response=response)
 
     return response.json()
 
@@ -71,6 +72,14 @@ def get_hail_variants_for_variant_ids(samples, genome_version, parsed_variant_id
         _validate_expected_families(response_json['results'], expected_family_guids)
 
     return response_json['results']
+
+
+def hail_variant_lookup(user, variant_id, genome_version=None, **kwargs):
+    return _execute_search({
+        'genome_version': GENOME_VERSION_LOOKUP[genome_version or GENOME_VERSION_GRCh38],
+        'variant_id': variant_id,
+        **kwargs,
+    }, user, path='lookup', exception_map={404: 'Variant not present in seqr'})
 
 
 def _format_search_body(samples, genome_version, num_results, search):