+++ seqr/views/apis/summary_data_api_tests.py
@@ -1,13 +1,14 @@
 from datetime import datetime
+from django.contrib.postgres.aggregates import ArrayAgg
 from django.urls.base import reverse
 import json
 import mock
 import responses
 
 from seqr.views.apis.summary_data_api import mme_details, success_story, saved_variants_page, hpo_summary_data, \
-    bulk_update_family_analysed_by, sample_metadata_export
-from seqr.views.utils.test_utils import AuthenticationTestCase, AnvilAuthenticationTestCase, AirtableTest
-from seqr.models import FamilyAnalysedBy
+    bulk_update_family_external_analysis, sample_metadata_export
+from seqr.views.utils.test_utils import AuthenticationTestCase, AnvilAuthenticationTestCase, AirtableTest, PARSED_VARIANTS
+from seqr.models import FamilyAnalysedBy, SavedVariant, VariantTag
 from settings import AIRTABLE_URL
 
 
@@ -75,7 +76,7 @@ EXPECTED_NO_AIRTABLE_SAMPLE_METADATA_ROW = {
     "family_id": "12",
     "MME": "Y",
     "subject_id": "NA20889",
-    "proband_relationship": "",
+    "proband_relationship": "Self",
     "consanguinity": "None suspected",
 }
 EXPECTED_SAMPLE_METADATA_ROW = {
@@ -85,6 +86,40 @@ EXPECTED_SAMPLE_METADATA_ROW = {
     "multiple_datasets": "No",
 }
 EXPECTED_SAMPLE_METADATA_ROW.update(EXPECTED_NO_AIRTABLE_SAMPLE_METADATA_ROW)
+EXPECTED_NO_GENE_SAMPLE_METADATA_ROW = {
+    'subject_id': 'NA21234',
+    'sample_id': 'NA21234',
+    'family_guid': 'F000014_14',
+    'family_id': '14',
+    'project_guid': 'R0004_non_analyst_project',
+    'project_id': 'Non-Analyst Project',
+    'affected_status': 'Affected',
+    'ancestry': '',
+    'congenital_status': 'Unknown',
+    'consanguinity': 'None suspected',
+    'data_type': 'WGS',
+    'date_data_generation': '2018-02-05',
+    'hpo_absent': '',
+    'hpo_present': '',
+    'inheritance_description-1': 'Autosomal dominant',
+    'maternal_id': '',
+    'MME': 'Y',
+    'novel_mendelian_gene-1': 'Y',
+    'num_saved_variants': 1,
+    'paternal_id': '',
+    'phenotype_description': None,
+    'phenotype_group': '',
+    'pmid_id': None,
+    'proband_relationship': 'Self',
+    'sex': 'Female',
+    'solve_state': 'Tier 1',
+    'Alt-1': 'T',
+    'Chrom-1': '1',
+    'Gene_Class-1': 'Tier 1 - Candidate',
+    'Pos-1': '248367227',
+    'Ref-1': 'TC',
+    'Zygosity-1': 'Heterozygous',
+}
 
 AIRTABLE_SAMPLE_RECORDS = {
   "records": [
@@ -248,8 +283,6 @@ class SummaryDataAPITest(AirtableTest):
         expected_variant_guids = {
             'SV0000001_2103343353_r0390_100', 'SV0000007_prefix_19107_DEL_r00', 'SV0000006_1248367227_r0003_tes',
         }
-        if self.MANAGER_VARIANT_GUID:
-            expected_variant_guids.add(self.MANAGER_VARIANT_GUID)
         self.assertSetEqual(set(response_json['savedVariantsByGuid'].keys()), expected_variant_guids)
         self.assertSetEqual(
             set(response_json['projectsByGuid'][PROJECT_GUID].keys()),
@@ -262,7 +295,6 @@ class SummaryDataAPITest(AirtableTest):
         self.assertEqual(response.status_code, 200)
         response_json = response.json()
         self.assertSetEqual(set(response_json.keys()), SAVED_VARIANT_RESPONSE_KEYS)
-        expected_variant_guids.discard(self.MANAGER_VARIANT_GUID)
         self.assertSetEqual(set(response_json['savedVariantsByGuid'].keys()), expected_variant_guids)
 
         all_tag_url = reverse(saved_variants_page, args=['ALL'])
@@ -348,14 +380,18 @@ class SummaryDataAPITest(AirtableTest):
             },
         ])
 
+    @mock.patch('seqr.views.apis.summary_data_api.datetime')
+    @mock.patch('seqr.views.apis.summary_data_api.get_variants_for_variant_ids')
     @mock.patch('seqr.views.apis.summary_data_api.load_uploaded_file')
-    def test_bulk_update_family_analysed_by(self, mock_load_uploaded_file):
-        url = reverse(bulk_update_family_analysed_by)
+    def test_bulk_update_family_external_analysis(self, mock_load_uploaded_file, mock_get_variants_for_variant_ids, mock_datetime):
+        mock_datetime.now.return_value = datetime(2023, 12, 5, 20, 16, 1)
+
+        url = reverse(bulk_update_family_external_analysis)
         self.check_analyst_login(url)
 
         mock_load_uploaded_file.return_value = [['foo', 'bar']]
-        response = self.client.post(url, content_type='application/json', data=json.dumps(
-            {'dataType': 'RNA', 'familiesFile': {'uploadedFileId': 'abc123'}}))
+        body = {'dataType': 'RNA', 'familiesFile': {'uploadedFileId': 'abc123'}}
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.json()['error'], 'Project and Family columns are required')
 
@@ -367,8 +403,7 @@ class SummaryDataAPITest(AirtableTest):
             ['not_a_project', '2'],
         ]
         created_time = datetime.now()
-        response = self.client.post(url, content_type='application/json', data=json.dumps(
-            {'dataType': 'RNA', 'familiesFile': {'uploadedFileId': 'abc123'}}))
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
         self.assertDictEqual(response.json(), {
             'warnings': [
                 'No match found for the following families: not_a_family (Test Reprocessed Project), 2 (not_a_project)'
@@ -381,6 +416,82 @@ class SummaryDataAPITest(AirtableTest):
         self.assertSetEqual({fab.data_type for fab in models}, {'RNA'})
         self.assertSetEqual({fab.created_by for fab in models}, {self.analyst_user})
         self.assertSetEqual({fab.family.family_id for fab in models}, {'1', '12'})
+
+        # Test AIP
+        aip_upload = {
+            'metadata': {
+                'categories': {
+                    '1': 'ClinVar Pathogenic',
+                    '2': 'New Gene-Disease Association',
+                    '3': 'High Impact Variant',
+                    '4': 'De-Novo',
+                    'support': 'High in Silico Scores'
+                }
+            },
+            'results': {
+                'HG00731': {
+                    '12-48367227-TC-T': {'categories': ['3', '4'], 'support_vars': ['2-103343353-GAGA-G']},
+                    '1-248367227-TC-T': {'categories': ['1'], 'support_vars': ['12-48367227-TC-T']},
+                },
+                'SAM_123': {
+                    '12-48367227-TC-T': {'categories': ['4', 'support'], 'support_vars': []},
+                },
+            }
+        }
+        mock_load_uploaded_file.return_value = aip_upload
+        mock_get_variants_for_variant_ids.return_value = PARSED_VARIANTS
+        body['dataType'] = 'AIP'
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['errors'], ['Unable to find the following individuals: SAM_123'])
+
+        aip_upload['results']['NA20889'] = aip_upload['results'].pop('SAM_123')
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()['errors'], [
+            "Unable to find the following family's AIP variants in the search backend: 2 (1-248367227-TC-T)",
+        ])
+
+        aip_upload['results']['HG00731']['2-103343353-GAGA-G'] = aip_upload['results']['HG00731'].pop('1-248367227-TC-T')
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.json(), {'info': ['Loaded 2 new and 1 updated AIP tags for 2 families']})
+
+        new_saved_variant = SavedVariant.objects.get(variant_id='2-103343353-GAGA-G')
+        self.assertDictEqual(new_saved_variant.saved_variant_json, PARSED_VARIANTS[1])
+
+        aip_tags = VariantTag.objects.filter(variant_tag_type__name='AIP').order_by('id').values(
+            'metadata', saved_variant_ids=ArrayAgg('saved_variants__id'))
+        self.assertEqual(len(aip_tags), 4)
+
+        existing_tag = aip_tags[0]
+        self.assertListEqual(existing_tag['saved_variant_ids'], [2])
+        self.assertDictEqual(
+            json.loads(existing_tag['metadata']), {
+                '3': {'name': 'High Impact Variant', 'date': '2023-12-05'},
+                '4': {'name': 'de Novo', 'date': '2023-11-15'},
+                'removed': {
+                    'support': {'date': '2023-11-15', 'name': 'High in Silico Scores'},
+                },
+            })
+
+        new_saved_variant_tag = aip_tags[2]
+        self.assertListEqual(new_saved_variant_tag['saved_variant_ids'], [new_saved_variant.id])
+        self.assertDictEqual(
+            json.loads(new_saved_variant_tag['metadata']),
+            {'1': {'name': 'ClinVar Pathogenic', 'date': '2023-12-05'}},
+        )
+
+        comp_het_tag = aip_tags[1]
+        self.assertSetEqual(set(comp_het_tag['saved_variant_ids']), {2, new_saved_variant.id})
+        self.assertIsNone(comp_het_tag['metadata'])
+
+        existing_variant_new_tag = aip_tags[3]
+        self.assertListEqual(existing_variant_new_tag['saved_variant_ids'], [6])
+        self.assertDictEqual(
+            json.loads(existing_variant_new_tag['metadata']),
+            {'4': {'name': 'De-Novo', 'date': '2023-12-05'}, 'support': {'name': 'High in Silico Scores', 'date': '2023-12-05'}},
+        )
 
         self.check_no_analyst_no_access(url)
 
@@ -437,10 +548,16 @@ class SummaryDataAPITest(AirtableTest):
         self.assertEqual(response.status_code, 403)
         self.assertEqual(response.json()['error'], 'Permission Denied')
 
+        # Test no gene for discovery variant
+        self.login_manager()
+        no_analyst_project_url = reverse(sample_metadata_export, args=['R0004_non_analyst_project'])
+        response = self.client.get(no_analyst_project_url)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.json(), {'rows': [EXPECTED_NO_GENE_SAMPLE_METADATA_ROW]})
+
         # Test analyst access
         self.login_analyst_user()
-        unauthorized_project_url = reverse(sample_metadata_export, args=['R0004_non_analyst_project'])
-        response = self.client.get(unauthorized_project_url)
+        response = self.client.get(no_analyst_project_url)
         self.assertEqual(response.status_code, 403)
         self.assertEqual(response.json()['error'], 'Permission Denied')
 
@@ -537,7 +654,6 @@ class SummaryDataAPITest(AirtableTest):
 class LocalSummaryDataAPITest(AuthenticationTestCase, SummaryDataAPITest):
     fixtures = ['users', '1kg_project', 'reference_data']
     NUM_MANAGER_SUBMISSIONS = 4
-    MANAGER_VARIANT_GUID = 'SV0000006_1248367227_r0004_non'
     ADDITIONAL_SAMPLES = ['NA21234']
 
 
@@ -554,7 +670,6 @@ def assert_has_expected_calls(self, users, skip_group_call_idxs=None):
 class AnvilSummaryDataAPITest(AnvilAuthenticationTestCase, SummaryDataAPITest):
     fixtures = ['users', 'social_auth', '1kg_project', 'reference_data']
     NUM_MANAGER_SUBMISSIONS = 4
-    MANAGER_VARIANT_GUID = 'SV0000006_1248367227_r0004_non'
     ADDITIONAL_SAMPLES = []
 
     def test_mme_details(self, *args):